Split into lib and bin
Enable testing lib components as explicit unittests
and using the lib in examples for integration type tests.
