+++ Cargo.toml
@@ -21,7 +21,3 @@ futures = "0.3.1"
 [dev-dependencies]
 serde_yaml = "0.8.11"
 tokio-test = "0.2.0-alpha.6"
-
-[[bin]]
-name = "gordo-controller"
-path = "src/main.rs"
