+++ src/crd/gordo/gordo.rs
@@ -84,7 +84,7 @@ impl Default for GordoSubmissionStatus {
 /// has set in its `metadata.generation`; meaning changes have been submitted to the resource but
 /// the `GordoStatus` has not been updated to reflect this and therefore needs to be submitted to
 /// the workflow generator job.
-pub(crate) async fn launch_waiting_gordo_workflows(
+pub async fn launch_waiting_gordo_workflows(
     resource: &Api<Gordo>,
     client: &APIClient,
     namespace: &str,
@@ -134,7 +134,7 @@ pub(crate) async fn launch_waiting_gordo_workflows(
 
 /// Start a gordo-deploy job using this `Gordo`.
 /// Will patch the status of the `Gordo` to reflect the current revision number
-pub(crate) async fn start_gordo_deploy_job(
+pub async fn start_gordo_deploy_job(
     gordo: &Gordo,
     client: &APIClient,
     resource: &Api<Gordo>,
@@ -179,7 +179,7 @@ pub(crate) async fn start_gordo_deploy_job(
 }
 
 /// Remove any gordo deploy jobs associated with this `Gordo`
-pub(crate) async fn remove_gordo_deploy_jobs(gordo: &Gordo, client: &APIClient, namespace: &str) -> () {
+pub async fn remove_gordo_deploy_jobs(gordo: &Gordo, client: &APIClient, namespace: &str) -> () {
     info!("Removing any gordo-deploy jobs for Gordo: '{}'", &gordo.metadata.name);
 
     let jobs = Api::v1Job(client.clone()).within(&namespace);
