+++ src/lib.rs
@@ -1,0 +1,27 @@
+use serde::Deserialize;
+
+pub mod crd;
+pub mod deploy_job;
+pub use crd::gordo::Gordo;
+pub use deploy_job::DeployJob;
+
+#[derive(Deserialize, Debug, Clone)]
+pub struct GordoEnvironmentConfig {
+    deploy_image: String,
+}
+impl Default for GordoEnvironmentConfig {
+    fn default() -> Self {
+        GordoEnvironmentConfig {
+            deploy_image: "auroradevacr.azurecr.io/gordo-infrastructure/gordo-deploy".to_owned(),
+        }
+    }
+}
+
+// Get a minor version from standard SemVer string
+pub fn minor_version(deploy_version: &str) -> Option<u32> {
+    deploy_version
+        .split('.')
+        .nth(1)
+        .map(|v| v.parse::<u32>().ok())
+        .unwrap_or(None)
+}
