+++ src/main.rs
@@ -1,27 +1,8 @@
 use futures::future::join;
 use kube::{client::APIClient, config};
 use log::error;
-use serde::Deserialize;
 
-mod crd;
-mod deploy_job;
-#[cfg(test)]
-mod tests;
-
-use crate::crd::gordo::Gordo;
-use crate::deploy_job::DeployJob;
-
-#[derive(Deserialize, Debug, Clone)]
-pub struct GordoEnvironmentConfig {
-    deploy_image: String,
-}
-impl Default for GordoEnvironmentConfig {
-    fn default() -> Self {
-        GordoEnvironmentConfig {
-            deploy_image: "auroradevacr.azurecr.io/gordo-infrastructure/gordo-deploy".to_owned(),
-        }
-    }
-}
+use gordo_controller::GordoEnvironmentConfig;
 
 #[tokio::main]
 async fn main() -> () {
@@ -41,17 +22,8 @@ async fn main() -> () {
     let client = APIClient::new(kube_config);
 
     join(
-        crate::crd::gordo::monitor_gordos(&client, &namespace, &env_config),
-        crate::crd::model::monitor_models(&client, &namespace, &env_config),
+        gordo_controller::crd::gordo::monitor_gordos(&client, &namespace, &env_config),
+        gordo_controller::crd::model::monitor_models(&client, &namespace, &env_config),
     )
     .await;
-}
-
-// Get a minor version from standard SemVer string
-pub fn minor_version(deploy_version: &str) -> Option<u32> {
-    deploy_version
-        .split('.')
-        .nth(1)
-        .map(|v| v.parse::<u32>().ok())
-        .unwrap_or(None)
 }
