--- src/tests/mod.rs
@@ -1,3 +1,0 @@
-pub(crate) mod helpers;
-#[cfg(test)]
-pub(crate) mod test_controller;
